release 2.9.1

+++ src/avt/IVP/MemStream.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -628,7 +628,7 @@ MemStream::LoadFile( const char *filename )
 
         CheckSize( _len );
         res = fread( _data, sizeof(_data[0]), _len, fp );
-        if (res != sizeof(_data[0])*_len)
+        if ((size_t)res != sizeof(_data[0])*_len)
         {
             debug1 << "Bad read of MemStream from " << filename << endl;
         }
