+++ src/plots/Molecule/avtMoleculeRenderer.h
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2014, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -101,7 +101,7 @@ class avtMoleculeRenderer : public avtCustomRenderer
     // lighting props
     virtual void GlobalLightingOn();
     virtual void GlobalLightingOff();
-    virtual void GlobalSetAmbientCoefficient(const float);
+    virtual void GlobalSetAmbientCoefficient(const double);
     virtual void SetSpecularProperties(bool,double,double,
                                        const ColorAttribute&);
